% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison_functions.R
\name{relate_strata}
\alias{relate_strata}
\title{Relate one set of strata to another}
\usage{
relate_strata(strata1, strata2)
}
\arguments{
\item{strata1}{A data frame of strata definitions. Usually this should be
the set with the smaller number of strata.}

\item{strata2}{A data frame of strata definitions. Usually this should be
the set with the larger number of strata.}
}
\value{
A named list, where element names are taken from \code{strata1}
  and each element is a vector of the indices of the related strata in
  \code{strata2}.
}
\description{
Takes two strata definition tables (e.g. StrataSpecht and StrataCERMB) and
returns a list where the number of elements is equal to the number of strata
in the first set (\code{strata1}), and each element is a vector with indices
for strata in the second set (\code{strata2}). See example below for more
detail.
}
\examples{
# Relate Specht strata to CERMB database strata (50cm vertical divisions)
# The resulting lookup list will have an element for each Specht stratum.
# Each element is a vector of integer indices of CERMB strata.
#
lookup <- relate_strata(CERMBlidar::StrataSpecht, CERMBlidar::StrataCERMB)

# Which 'StrataCERMB' strata relate to the second (low shrub) layer
# in 'StrataSpecht'?
lookup[[2]]

}
